--- a/.eslintrc
+++ b/.eslintrc
@@ -1,5 +1,5 @@
 {
-  "extends": "eslint:recommended",
+  "extends": "xo",
   "env": {
     "node": true,
     "es6": true