--- a/index.js
+++ b/index.js
@@ -1,47 +1,47 @@
-'use strict'
+'use strict';
 
 /**
  * Helpers.
  */
 
-var s = 1000
-var m = s * 60
-var h = m * 60
-var d = h * 24
-var y = d * 365.25
+var s = 1000;
+var m = s * 60;
+var h = m * 60;
+var d = h * 24;
+var y = d * 365.25;
 
 var scales = {
-  years: y,
-  year: y,
-  yrs: y,
-  yr: y,
-  y: y,
-  days: d,
-  day: d,
-  d: d,
-  hours: h,
-  hour: h,
-  hrs: h,
-  hr: h,
-  h: h,
-  minutes: m,
-  minute: m,
-  mins: m,
-  min: m,
-  m: m,
-  seconds: s,
-  second: s,
-  secs: s,
-  sec: s,
-  s: s,
-  milliseconds: 1,
-  millisecond: 1,
-  msecs: 1,
-  msec: 1,
-  ms: 1
-}
+	years: y,
+	year: y,
+	yrs: y,
+	yr: y,
+	y: y,
+	days: d,
+	day: d,
+	d: d,
+	hours: h,
+	hour: h,
+	hrs: h,
+	hr: h,
+	h: h,
+	minutes: m,
+	minute: m,
+	mins: m,
+	min: m,
+	m: m,
+	seconds: s,
+	second: s,
+	secs: s,
+	sec: s,
+	s: s,
+	milliseconds: 1,
+	millisecond: 1,
+	msecs: 1,
+	msec: 1,
+	ms: 1
+};
 
-var splitExp = /\s*?(?=[a-z])/i
+var splitExp = /\s*?(?=[a-z])/i;
 
 /**
  * Parse the given `str` and return milliseconds.
@@ -50,27 +50,27 @@ var splitExp = /\s*?(?=[a-z])/i
  * @return {Number}
  * @api private
  */
-function parse (str) {
-  str = String(str)
-  var match = str.match(splitExp)
-  var numStr = str
-  var scale = ''
-  if (match) {
-    var scalePos = match.index + match[0].length
-    if (scalePos < numStr.length) {
-      scale = numStr.substr(scalePos).toLowerCase()
-      numStr = numStr.substr(0, scalePos)
-    }
-  }
-  var num = parseFloat(+numStr)
-  if (!isFinite(num)) {
-    return
-  }
-  scale = scale || 'ms'
-  if (!(scale in scales)) {
-    return
-  }
-  return num * scales[scale]
+function parse(str) {
+	str = String(str);
+	var match = str.match(splitExp);
+	var numStr = str;
+	var scale = '';
+	if (match) {
+		var scalePos = match.index + match[0].length;
+		if (scalePos < numStr.length) {
+			scale = numStr.substr(scalePos).toLowerCase();
+			numStr = numStr.substr(0, scalePos);
+		}
+	}
+	var num = parseFloat(Number(numStr));
+	if (!isFinite(num)) {
+		return;
+	}
+	scale = scale || 'ms';
+	if (!(scale in scales)) {
+		return;
+	}
+	return num * scales[scale];
 }
 
 /**
@@ -81,20 +81,20 @@ function parse (str) {
  * @api private
  */
 
-function fmtShort (ms) {
-  if (ms >= d) {
-    return Math.round(ms / d) + 'd'
-  }
-  if (ms >= h) {
-    return Math.round(ms / h) + 'h'
-  }
-  if (ms >= m) {
-    return Math.round(ms / m) + 'm'
-  }
-  if (ms >= s) {
-    return Math.round(ms / s) + 's'
-  }
-  return ms + 'ms'
+function fmtShort(ms) {
+	if (ms >= d) {
+		return Math.round(ms / d) + 'd';
+	}
+	if (ms >= h) {
+		return Math.round(ms / h) + 'h';
+	}
+	if (ms >= m) {
+		return Math.round(ms / m) + 'm';
+	}
+	if (ms >= s) {
+		return Math.round(ms / s) + 's';
+	}
+	return ms + 'ms';
 }
 
 /**
@@ -105,24 +105,24 @@ function fmtShort (ms) {
  * @api private
  */
 
-function fmtLong (ms) {
-  return (
+function fmtLong(ms) {
+	return (
     plural(ms, d, 'day') || plural(ms, h, 'hour') || plural(ms, m, 'minute') || plural(ms, s, 'second') || ms + ' ms'
-  )
+	);
 }
 
 /**
  * Pluralization helper.
  */
 
-function plural (ms, n, name) {
-  if (ms < n) {
-    return
-  }
-  if (ms < n * 1.5) {
-    return Math.floor(ms / n) + ' ' + name
-  }
-  return Math.ceil(ms / n) + ' ' + name + 's'
+function plural(ms, n, name) {
+	if (ms < n) {
+		return;
+	}
+	if (ms < n * 1.5) {
+		return Math.floor(ms / n) + ' ' + name;
+	}
+	return Math.ceil(ms / n) + ' ' + name + 's';
 }
 
 /**
@@ -140,12 +140,12 @@ function plural (ms, n, name) {
  */
 
 module.exports = function (val, options) {
-  options = options || {}
-  var type = typeof val
-  if (type === 'string' && val.length > 0) {
-    return parse(val)
-  } else if (type === 'number' && isFinite(val)) {
-    return options.long ? fmtLong(val) : fmtShort(val)
-  }
-  throw new Error('val is not a non-empty string or a valid number. val=' + JSON.stringify(val))
-}
+	options = options || {};
+	var type = typeof val;
+	if (type === 'string' && val.length > 0) {
+		return parse(val);
+	} else if (type === 'number' && isFinite(val)) {
+		return options.long ? fmtLong(val) : fmtShort(val);
+	}
+	throw new Error('val is not a non-empty string or a valid number. val=' + JSON.stringify(val));
+};