--- a/package.json
+++ b/package.json
@@ -28,10 +28,11 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "eslint": "3.19.0",
     "expect.js": "0.3.1",
     "husky": "0.13.3",
     "lint-staged": "3.4.1",
-    "mocha": "3.4.1"
+    "mocha": "3.4.1",
+    "eslint": "3.19.0",
+    "prettier": "1.3.1"
   }
 }