--- a/package.json
+++ b/package.json
@@ -9,20 +9,13 @@
   ],
   "scripts": {
     "precommit": "lint-staged",
-    "lint": "eslint lib/* bin/*",
+    "lint": "eslint *.js",
     "test": "mocha tests.js"
-  },
-  "eslintConfig": {
-    "extends": "eslint:recommended",
-    "env": {
-      "node": true,
-      "es6": true
-    }
   },
   "lint-staged": {
     "*.js": [
+      "standard --fix *.js",
       "npm run lint",
-      "prettier --single-quote --write",
       "git add"
     ]
   },
@@ -33,6 +26,6 @@
     "lint-staged": "3.4.1",
     "mocha": "3.4.1",
     "eslint": "3.19.0",
-    "prettier": "1.3.1"
+    "standard": "10.0.2"
   }
 }