--- a/package.json
+++ b/package.json
@@ -9,12 +9,11 @@
   ],
   "scripts": {
     "precommit": "lint-staged",
-    "lint": "eslint *.js",
+    "lint": "eslint --fix *.js",
     "test": "mocha tests.js"
   },
   "lint-staged": {
     "*.js": [
-      "standard --fix *.js",
       "npm run lint",
       "git add"
     ]
@@ -22,10 +21,9 @@
   "license": "MIT",
   "devDependencies": {
     "expect.js": "0.3.1",
-    "husky": "0.13.3",
     "lint-staged": "3.4.1",
     "mocha": "3.4.1",
     "eslint": "3.19.0",
-    "standard": "10.0.2"
+    "eslint-config-xo": "0.18.2"
   }
 }