- dramatically performance optimization
- added test for Infinity and long strings performance
- standard style fix, fix missing devDependencies