@leo 2017-05-17
Looks good to me! But can you please remove `standard` and replace it with `xo`? Or simply leave it out and we'll add it later.