@Delagen 2017-05-17
What you don't like in standard? May be left only eslint with fixers!?