@Delagen 2017-05-17
replaced by eslint-config-xo which IMHO better than second linter