@thevtm 2017-06-07
Hi!
I did a little benchmarking and found out that the original implementation seems to be around 2x faster than the PR implementation.

[Benchmark source](https://gist.github.com/thevtm/063a834083a42af535e26196a5d3c8d5)

```
original x 350,875 ops/sec ±1.68% (84 runs sampled)
new x 130,267 ops/sec ±1.02% (87 runs sampled)
Fastest is original
Done in 12.50s.
```