@Delagen 2017-06-07
@thevtm Just insert one test value `'1'.repeat(29) + 'Z'` to invert results