@thevtm 2017-06-07
@Delagen thanks! I haven't considered that case.
I ended up finding a better solution.
Changing the Regex from:
 `/^((?:\d+)?\.?\d+) *(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|years?|yrs?|y)?$/i`
to:
`/^((?:\d+)?\.?\d+) *(\w{0,12})?$/i`
seems to produce better results in all cases.

[Benchmark source](https://gist.github.com/thevtm/063a834083a42af535e26196a5d3c8d5)

```
original x 38,472 ops/sec ±2.05% (88 runs sampled)
new x 39,713 ops/sec ±1.89% (82 runs sampled)
mine x 74,820 ops/sec ±1.84% (84 runs sampled)
Fastest is mine
Done in 17.97s.
```