@Delagen 2017-06-07
@thevtm thanks for benchmarking, updated code with your regexp