@Delagen 2017-12-01
@TooTallNate if you about IE8, It's dead, even webpack removed support of it