@ForbesLindesay 2018-06-01
@Delagen Even IE10 doesn't support Map. IE11's support is pretty hit and miss (although it probably supports everything you need here): https://kangax.github.io/compat-table/es6/