@Delagen 2018-06-01
@ForbesLindesay Modern development use a couple bunch of polyfills or transpilers when target IE, so it's not a problem at all. This PR is about 1 year old, so it can never merge.