@rauchg 2018-12-03
Interested in this, but PR has to be re-factored